requirements.txt
# Azure AI Foundry SDK
azure-ai-projects
azure-identity
azure-ai-inference
azure-keyvault-secrets

# Agent Framework
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0

# Database - PostgreSQL
psycopg[binary]>=3.1.0
sqlalchemy>=2.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Additional utilities
requests>=2.31.0
aiohttp>=3.9.0
